Added testGenerate goal to rosetta-maven-plugin
