+++ rosetta-maven-plugin/src/main/java/com/regnosys/rosetta/maven/RosettaGenerateMojo.java
@@ -1,123 +1,57 @@
 package com.regnosys.rosetta.maven;
 
-import java.io.File;
-import java.util.List;
-import java.util.Map;
+import static com.google.common.collect.Iterables.filter;
+import static com.google.common.collect.Sets.newLinkedHashSet;
 
-import org.apache.maven.plugin.MojoExecutionException;
+import java.util.List;
+import java.util.Set;
+
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.plugins.annotations.ResolutionScope;
-import org.eclipse.xtext.builder.standalone.LanguageAccess;
-import org.eclipse.xtext.builder.standalone.LanguageAccessFactory;
-import org.eclipse.xtext.builder.standalone.StandaloneBuilder;
-import org.eclipse.xtext.builder.standalone.compiler.CompilerConfiguration;
-import org.eclipse.xtext.builder.standalone.compiler.IJavaCompiler;
-import org.eclipse.xtext.maven.ClusteringConfig;
-import org.eclipse.xtext.maven.XtextGenerateMojo;
-import org.eclipse.xtext.xbase.lib.IterableExtensions;
+import org.eclipse.xtext.maven.Language;
 
-import com.google.inject.Guice;
-import com.google.inject.Injector;
-import com.google.inject.Module;
 
-// TODO: decouple from `XtextGenerateMojo`/`xtext-maven-plugin`. They differ too much,
-// which makes this implementation too hacky.
+// This class is identical to `XtextGenerateMojo`, except its superclass.
 @Mojo(name = "generate", defaultPhase = LifecyclePhase.GENERATE_SOURCES, requiresDependencyResolution = ResolutionScope.COMPILE, threadSafe = true)
-public class RosettaGenerateMojo extends XtextGenerateMojo {
+public class RosettaGenerateMojo extends AbstractRosettaGeneratorMojo {
 
-	@Parameter(defaultValue = "true")
-	boolean addOutputDirectoriesToCompileSourceRoots = Boolean.TRUE;
-	
-	@Parameter
-	private String classPathLookupFilter;
-	
-	@Parameter(defaultValue = "${project.compileSourceRoots}", required = true)
-	private List<String> javaSourceRoots;
-	
-	@Parameter(defaultValue = "true")
-	private Boolean failOnValidationError;
-	
-	@Parameter(defaultValue = "${project.build.directory}/xtext-temp")
-	private String tmpClassDirectory;
-	
-	@Parameter
-	private ClusteringConfig clusteringConfig;
-	
-	@Parameter(property = "maven.compiler.source", defaultValue = "1.6")
-	private String compilerSourceLevel;
+	/**
+	 * Project classpath.
+	 */
+	@Parameter(defaultValue = "${project.compileClasspathElements}", readonly = true, required = true)
+	private List<String> classpathElements;
 
-	@Parameter(property = "maven.compiler.target", defaultValue = "1.6")
-	private String compilerTargetLevel;
-	
-	@Parameter(defaultValue = "false")
-	private Boolean compilerSkipAnnotationProcessing;
-
-	@Parameter(defaultValue = "false")
-	private Boolean compilerPreserveInformationAboutFormalParameters;
-	
-	// TODO: add this method to Xtext so I don't have to overwrite `internalExecute`
-	// and duplicate all of the above parameters.
-	protected Module createModule() {
-		return new RosettaMavenStandaloneBuilderModule();
-	}
-	
 	@Override
-	protected void internalExecute() throws MojoExecutionException {
-		if (addOutputDirectoriesToCompileSourceRoots) {
-			configureMavenOutputs();
-		}
-		Map<String, LanguageAccess> languages = new LanguageAccessFactory().createLanguageAccess(getLanguages(),
-				this.getClass().getClassLoader());
-		Injector injector = Guice.createInjector(createModule());
-		StandaloneBuilder builder = injector.getInstance(StandaloneBuilder.class);
-		builder.setBaseDir(getProject().getBasedir().getAbsolutePath());
-		builder.setLanguages(languages);
-		builder.setEncoding(getEncoding());
-		builder.setClassPathEntries(getClasspathElements());
-		builder.setClassPathLookUpFilter(classPathLookupFilter);
-		builder.setSourceDirs(getSourceRoots());
-		builder.setJavaSourceDirs(javaSourceRoots);
-		builder.setFailOnValidationError(failOnValidationError);
-		builder.setTempDir(createTempDir().getAbsolutePath());
-		builder.setDebugLog(getLog().isDebugEnabled());
-		if (clusteringConfig != null)
-			builder.setClusteringConfig(clusteringConfig.convertToStandaloneConfig());
-		configureCompiler(builder.getCompiler());
-		logState();
-		boolean errorDetected = !builder.launch();
-		if (errorDetected && failOnValidationError) {
-			throw new MojoExecutionException("Execution failed due to a severe validation error.");
-		}
-	}
-	
-	private void configureCompiler(IJavaCompiler compiler) {
-		CompilerConfiguration conf = compiler.getConfiguration();
-		conf.setSourceLevel(compilerSourceLevel);
-		conf.setTargetLevel(compilerTargetLevel);
-		conf.setVerbose(getLog().isDebugEnabled());
-		conf.setSkipAnnotationProcessing(compilerSkipAnnotationProcessing);
-		conf.setPreserveInformationAboutFormalParameters(compilerPreserveInformationAboutFormalParameters);
+	public Set<String> getClasspathElements() {
+		Set<String> classpathElements = newLinkedHashSet();
+		classpathElements.addAll(this.classpathElements);
+		classpathElements.remove(getProject().getBuild().getOutputDirectory());
+		classpathElements.remove(getProject().getBuild().getTestOutputDirectory());
+		Set<String> nonEmptyElements = newLinkedHashSet(filter(classpathElements, emptyStringFilter()));
+		return nonEmptyElements;
 	}
 
-	private File createTempDir() {
-		File tmpDir = new File(tmpClassDirectory);
-		if (!tmpDir.mkdirs() && !tmpDir.exists()) {
-			throw new IllegalArgumentException("Couldn't create directory '" + tmpClassDirectory + "'.");
+	@Override
+	protected void configureMavenOutputs() {
+		for (Language language : getLanguages()) {
+			addCompileSourceRoots(language);
 		}
-		return tmpDir;
 	}
 	
-	private void logState() {
-		getLog().info(
-				"Encoding: " + (getEncoding() == null ? "not set. Encoding provider will be used." : getEncoding()));
-		getLog().info("Compiler source level: " + compilerSourceLevel);
-		getLog().info("Compiler target level: " + compilerTargetLevel);
-		if (getLog().isDebugEnabled()) {
-			getLog().debug("Source dirs: " + IterableExtensions.join(getSourceRoots(), ", "));
-			getLog().debug("Java source dirs: " + IterableExtensions.join(javaSourceRoots, ", "));
-			getLog().debug("Classpath entries: " + IterableExtensions.join(getClasspathElements(), ", "));
-		}
+	/**
+	 * Project source roots. List of folders, where the source models are
+	 * located.<br>
+	 * The default value is a reference to the project's
+	 * ${project.compileSourceRoots}.<br>
+	 * When adding a new entry the default value will be overwritten not extended.
+	 */
+	@Parameter(defaultValue = "${project.compileSourceRoots}", required = true)
+	private List<String> sourceRoots;
+
+	@Override
+	protected List<String> getSourceRoots() {
+		return sourceRoots;
 	}
 }
